MABEL v2.10 Datatype Description and Example
Sometimes, Python may infer incorrect datatypes for certain features or values in the MABEL dataset. Most features in MABEL are represented as either int64 or object (string) types.

This datatype description file provides a reference for the correct datatype of each feature in the dataset.
After loading the MABEL dataset, you can use this file to accurately set the datatypes for all features.

For clarity, the file also includes an “example” column showing a sample value for each feature. This helps illustrate the expected format and assists in validation or debugging during data preprocessing.